requirements.txt
requests
python-dotenv
sqlite3
datetime
